+++ CHANGELOG.md
@@ -6,3 +6,7 @@
 
 # 1.0.6
 - Add auto ip selection support
+
+# 1.0.7
+- Fix udp server address problem
+- Support config path: /etc/shadowsocks-lite/config.json